--- src/app/(provider)/(root)/_components/CategorySection/index.ts
@@ -1,2 +1,0 @@
-export { default as CategoryProducts } from './CategoryProducts';
-export { default as CategorySection } from './CategorySection';
