--- src/app/(provider)/(root)/_components/EventLinkCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./EventLinkCard";
