+++ src/app/(provider)/(root)/products/_components/DetailButtons/index.ts
@@ -1,2 +1,3 @@
+export { default as Paying } from './Paying';
 export { default as Share } from './Share';
 export { default as Wish } from './Wish';
