--- src/app/(provider)/(root)/products/_components/Paying/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Paying';
