--- src/components/BannerSlide/index.ts
@@ -1,2 +1,0 @@
-export { default as BannerSlide } from './BannerSlide';
-export { default as ProductSlide } from './ProductSlide';
