--- src/components/ReviewCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ReviewCard';
