--- src/components/SkeletonCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SkeletonCard';
