+++ src/components/Sliders/CategoryProducts.tsx
@@ -1,6 +1,6 @@
 'use client';
 import CategoryMore from '@/components/CategoryMore';
-import Sliders from '@/components/Sliders';
+import { Sliders } from '@/components/Sliders';
 import useProductsQuery from '@/hooks/query/useProductsQuery';
 import Image from 'next/image';
