+++ src/components/Sliders/index.ts
@@ -1,1 +1,7 @@
-export { default } from './Sliders';
+export { default as BannerSlide } from './BannerSlide';
+export { default as CategoryProducts } from './CategoryProducts';
+export { default as CategorySection } from './CategorySection';
+export { default as CurrentProducts } from './CurrentProducts';
+export { default as MdReviews } from './MdReviews';
+export { default as ProductSlide } from './ProductSlide';
+export { default as Sliders } from './Sliders';
